Update Password Controller
Remove the placeholder message and call the resetPassword message on the $user object to check whether the password is ok.

If the password is ok, then echo a generic message for now.

If the password is not ok, then render the Password/reset View passing along the token and user object.