Create Password/reset_success View
Create a new View which simply displays a success message and a link to login.

This will only be displayed if the insertion query for the new password is successful.

It is not necessary to keep the errors array, because the insertion query in the User model will not even attempt to run if there is anything in the errors array.