Update User.php
Update comments to /**.

Change bindParam to bindValue.

bindParam binds a reference, whereas bindValue binds a value.

Since user_email is a unique index in the database, and we are only searching for one record in the database in this method, bindValue should be fine.

This also makes the User model in my codebase match Dave H.'s.