Update Profile/edit View
Remove password as a requirement.

Add span with text to explain that the password is not required; leaving it blank will keep the existing password.

Add id to the span.

Reference the span id while adding an aria-describedby attribute to the password input - this lets a screen-reader know that the span provides a description for the <input>