Add Action Filter to Profile Controller
Add a before action filter to the Profile Controller to get the $user object.

Make sure to call the parent before action filter, or else the child will override the parent.

parent is a keyword in PHP, and in this case, it refers to the Authenticated class, which the Profile class extends.