Edit Password/requestResetAction
Before, after requesting the password reset email, the Flash would not display on the confirmation page. This is because the Flash message only displayed after redirect.

To remedy this, move the confirmation render to a new method (resetConfirmation), and redirect to this page after sending the reset email and setting the Flash message.