Add 2 vars and 2 methods
Add a var for email address and for the Mailer object. Use these new variables in the setMailer method, which accepts a Mailer object as an argument, and stores that argument at the mailer index of the current object. This mailer object is actually set within the test as a test double, so it is actually a mock mailer. The notify method accepts a  as the argument and simply returns the result from using the sendMessage method of the current object's mailer index (the injected mock Mailer). Sorry if this made no sense, PHPUnit doesn't make much sense to me either