Add 2 assertions to testNotificationIsSent
Add 2 assertions to the testNotificationIsSent method: 1. expects and 2. with. Expects takes (->once()) as its argument, which tests whether the stubbed method is run exactly 1 time; could also use (->never()) or etc. The assertion with() takes expected values for the sendMessage stubbed arguments, the first of which is the , and the second of which is the . So test whether the  is equal to dave@example.com.